backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.13.1